Move middleware to API package
